v1(for local use)/requirements.txt
httpx
python-dotenv
pandas
pypdf2